==> Mundial2022/2.LimpiarDatosMundial.ipynb <==
import pandas as pd

df_data_historica = pd.read_csv('fifa_worldcup_historical_data.csv')
df_fixture = pd.read_csv('fifa_worldcup_fixture.csv')
df_fixture

df_fixture['home'].str.strip()
df_fixture['away'].str.strip()

#Eliminamos este partido porque fue dado a Suecia por 3-0
index_eliminar = df_data_historica[df_data_historica['home'].str.contains('Sweden') & df_data_historica['away'].str.contains('Austria')].index

df_data_historica.drop(index=index_eliminar, inplace=True)
df_data_historica

#Separamos la columna score en score_home y score_away
df_data_historica['score'].str.split('–', expand=True)

df_data_historica
